+++ src/pages/contract.tsx
@@ -1,54 +1,14 @@
 import { useWallet } from "@suiet/wallet-kit";
-import { useEffect, useState } from "react";
-import { SUI_PACKAGE, SUI_MODULE, NETWORK } from "../config/constants";
 import { JsonRpcProvider, TransactionBlock } from '@mysten/sui.js';
+import { useEffect, useState } from "react";
+import { SUI_PACKAGE, NETWORK } from "../config/constants";
 import TransacitonLink from "../utils/tools";
-
-type SwordListPros = { swords: Array<{ id: string, magic: number, strength: number }>, transfer: Function };
-const SwordList = ({ swords, transfer }: SwordListPros) => {
-    return swords && (
-        <div className="card lg:card-side bg-base-100 shadow-xl mt-5">
-            <div className="card-body">
-                <h2 className="card-title">swords list:</h2>
-
-                <div className="overflow-x-auto">
-                    <table className="table w-full">
-                        <thead>
-                            <tr>
-                                <th>Id</th>
-                                <th>Magic</th>
-                                <th>Strength</th>
-                                <th>Operate</th>
-                            </tr>
-                        </thead>
-                        <tbody>
-
-                            {
-                                swords.map((item, i) =>
-                                    <tr key={item.id}>
-                                        <th>{item.id}</th>
-                                        <td>{item.magic}</td>
-                                        <td>{item.strength} </td>
-                                        <td>
-                                            <a className="link link-hover link-primary" onClick={() => { transfer(item.id) }}>Transfer</a>
-                                        </td>
-                                    </tr>
-                                )
-                            }
-                        </tbody>
-                    </table>
-                </div>
-            </div>
-        </div>
-    )
-}
+import { SwordList } from "../components/SwordList";
 
 export default function Contract() {
     const [swords, setSwords] = useState<Array<any>>([]);
     const [transaction, setTransaction] = useState('');
-    const { connected, account, signAndExecuteTransactionBlock } = useWallet();
     const [displayModal, toggleDisplay] = useState(false);
-    const [message, setMessage] = useState('');
     const [sword_id, updateSwordId] = useState('')
     const [to_recipient, updateToRecipient] = useState("");
     const [formInput, updateFormInput] = useState<{
@@ -60,6 +20,7 @@ export default function Contract() {
         strength: "",
         recipient: "",
     });
+    const { connected, account, signAndExecuteTransactionBlock } = useWallet();
 
     async function transferSword(id: string) {
         console.log("transfer sword ", id);
@@ -68,22 +29,6 @@ export default function Contract() {
     }
 
     async function doTransfer() {
-        function makeTranscaction() {
-            return {
-                packageObjectId: SUI_PACKAGE,
-                module: SUI_MODULE,
-                function: 'sword_transfer',
-                typeArguments: [],
-                // 类型错误，传递字符串类型，部分钱包会内部转化
-                arguments: [
-                    sword_id,
-                    to_recipient,
-                ],
-                gasBudget: 30000,
-            };
-        }
-
-        setMessage("");
         try {
             const tx = new TransactionBlock();
             tx.moveCall({
@@ -100,13 +45,11 @@ export default function Contract() {
 
             updateFormInput({ magic: "", strength: "", recipient: "" })
             console.log('success', resData);
-            setMessage('Added succeeded');
             if (resData && resData.digest && resData.digest) {
                 setTransaction(TransacitonLink(resData.digest, "my_module"));
             }
         } catch (e) {
             console.error('failed', e);
-            setMessage('Mint failed: ' + e);
             setTransaction('');
         }
     }
@@ -128,7 +71,6 @@ export default function Contract() {
             }
         })
         console.log(objects);
-
         let swords = objects.data.map(item => {
             console.log("old item : ", item);
             if (item.data && item.data.content) {
@@ -155,7 +97,6 @@ export default function Contract() {
     }, [connected, transaction])
 
     const provider = new JsonRpcProvider();
-
     const createSword = async () => {
         try {
             const tx = new TransactionBlock();
@@ -175,7 +116,6 @@ export default function Contract() {
 
             updateFormInput({ magic: "", strength: "", recipient: "" })
             console.log('success', resData);
-            setMessage('Added succeeded');
             if (resData && resData.digest && resData.digest) {
                 setTransaction(TransacitonLink(resData.digest, "my_module"));
             }
@@ -214,7 +154,7 @@ export default function Contract() {
                 <div>
                     <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" className="stroke-current flex-shrink-0 w-6 h-6"><path strokeLinecap="round" strokeLinejoin="round" strokeWidth="2" d="M13 16h-1v-4h-1m1-4h.01M21 12a9 9 0 11-18 0 9 9 0 0118 0z"></path></svg>
                     <span>
-                        Sui Test Package Module: <a className="link link-primary" target={"_blank"} href={"https://explorer.sui.io/object/" + SUI_PACKAGE + "?network=" + NETWORK}>{SUI_PACKAGE}</a>
+                        Sui Package Module: <a className="link link-primary" target={"_blank"} href={"https://explorer.sui.io/object/" + SUI_PACKAGE + "?network=" + NETWORK}>{SUI_PACKAGE}</a>
                     </span>
                 </div>
             </div>
